Seperate addfolderstoPath into utils
